requirements.txt
requests==2.22.0
PyNaCl==1.3.0
google-cloud-firestore==1.7.0
firebase-admin==4.3.0
Quart==0.12.0
Quart-CORS==0.3.0
Hypercorn==0.10.1
